Removed cyclic dependencies. The header updater is no longer a stage. Instead it is written as part of the Image.writeto method.
